test(server): retire the routes/canvas.test.ts monolith into per-route files
routes/canvas.test.ts predated the routes/canvas/ split and had grown to
2323 lines / 77 tests overlapping the per-route test files it split from.
route-coverage.test.ts requires canvas.ts to keep a sibling .test.ts, so
canvas.test.ts survives as a ~150-line composition test holding only the
two cross-family integration scenarios (nested-path routing across every
sub-router) that exercise canvas.ts's own job of composing the sub-routers.

Zero drops: every other test moved verbatim (whole describe blocks, in
original order) into its owning per-route file below, with each file's
existing harness (withTempDataDir + config.js mock, doc-cache vi.fn wrap
only where interleaving control is used) replicated to receive it. None of
restore.test.ts's existing kind-propagation/node-count tests were a strict
assertion superset of their monolith counterparts (different projections:
list `kind` field vs direct getDocumentKind(), elementCount only vs also
documentId/broadcastCalls/peekDoc identity), so per the "move, don't drop
on uncertainty" rule those were moved too rather than treated as covered.

Moved:
- GET /api/workspaces, POST/DELETE/GET .../canvases, PUT .../path (rename,
  incl. the rename-vs-update phantom-duplicate race and both doc-cache-
  eviction tests) -> canvas/workspaces.test.ts (28 tests)
- GET .../snapshot, GET .../exists, POST .../update -> canvas/live-doc.test.ts
  (10 tests)
- POST .../compact, POST .../optimize-all (incl. prune-sandwiched)
  -> canvas/maintenance.test.ts (4 tests)
- versions API save/list/operator-stamping tests -> canvas/versions.test.ts
  (6 tests)
- in-place restore (incl. both evict-on-failure tests, 404/400) and the
  'overwrite restore reconciles instead of replacing' describe (targetPath
  handling, kind propagation, output_exists) -> canvas/restore.test.ts
  (12 tests)
- thumbnail PUT/GET/latest-thumbnail tests -> canvas/thumbnails.test.ts
  (8 tests)
- createAutoVersionTrigger + auto-version corruption handling
  -> canvas/auto-version.test.ts (7 tests)

Also fixes the stale stryker exclusion naming the monolith and its dead
comment ("Incomplete server-mode auth route" — no such test existed), and
updates three "mirroring canvas.test.ts's phantom-duplicate pin" comment
pointers (ws-rename-race.test.ts, app.merge-race.test.ts,
canvas/restore-race.test.ts) to canvas/workspaces.test.ts where that pin
now lives.

Verified via `pnpm exec vitest run --project mcp-node routes/canvas`:
119 tests before and after (77 monolith + 42 pre-existing canvas/* tests
before; 2 composition + 117 redistributed after) — no coverage lost. Full
`pnpm test --project mcp-node`: 3458 passed + 6 skipped (3464) before and
after. `pnpm exec vitest run -c packages/mcp-server/vitest.stryker.config.ts`
now includes the un-excluded tests and passes (256 files, 3229 passed + 5
skipped).
