+++ packages/mcp-server/src/server/app.merge-race.test.ts
@@ -29,7 +29,7 @@ vi.mock('../daemon/ensure-daemon.js', () => ({
 }))
 
 // Mock doc-cache so getDoc can be gated to control interleaving against a
-// concurrent rename, mirroring canvas.test.ts's phantom-duplicate pin.
+// concurrent rename, mirroring canvas/workspaces.test.ts's phantom-duplicate pin.
 vi.mock('./store/doc-cache.js', async () => {
   const actual =
     await vi.importActual<typeof import('./store/doc-cache.js')>('./store/doc-cache.js')
