+++ packages/mcp-server/src/server/routes/canvas.test.ts
@@ -1,13 +1,4 @@
-import { mkdir, writeFile } from 'node:fs/promises'
-import { userInfo } from 'node:os'
-import { join } from 'node:path'
-import { LoroDoc, LoroMap } from 'loro-crdt'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import {
-  deleteCanvasResponseSchema,
-  renameDocumentPathResponseSchema,
-  updateCanvasResponseSchema,
-} from '../../shared/api-contracts/canvas.js'
 import { withTempDataDir } from './_test-helpers.js'
 
 const tmp = withTempDataDir('whiteboard-routes-test-')
@@ -21,138 +12,24 @@ vi.mock('../config.js', () => ({
   REPO_ROOT: '/tmp',
 }))
 
-// Mock doc-cache so the cache is isolated in tests. getDoc is wrapped in
-// vi.fn (defaulting to the real implementation) so individual tests can
-// install a one-shot mockImplementationOnce to control interleaving against
-// a concurrent request.
-vi.mock('../store/doc-cache.js', async () => {
-  const actual =
-    await vi.importActual<typeof import('../store/doc-cache.js')>('../store/doc-cache.js')
-  return { ...actual, getDoc: vi.fn(actual.getDoc) }
-})
-
-const { clearCache, peekDoc, getDoc } = await import('../store/doc-cache.js')
-const { saveDocument } = await import('../store/document-store.js')
-const { corruptStoredData } = await import('../store/corrupt-stored-data.js')
-const { setBroadcastFn } = await import('./canvas/_shared.js')
+const { clearCache } = await import('../store/doc-cache.js')
 
 // Dynamically import the Hono app.
-const { createCanvasRouter, createAutoVersionTrigger } = await import('./canvas.js')
+const { createCanvasRouter } = await import('./canvas.js')
 
-describe('GET /api/workspaces', () => {
+// canvas.ts's own job is composing the canvas/*.ts sub-routers into one
+// router (see route-coverage.test.ts, which requires every route source
+// file to keep a sibling .test.ts). Per-behavior coverage for each
+// sub-router lives beside it under canvas/*.test.ts; this file keeps only
+// the two cross-family integration scenarios that exercise the composition
+// itself — a nested document path routed correctly across every sub-router
+// in one request sequence.
+describe('createCanvasRouter composition', () => {
   beforeEach(() => {
     clearCache()
   })
   afterEach(() => {
     clearCache()
-  })
-
-  it('returns the canonical workspace list with workspaceId entries', async () => {
-    await saveDocument('workspace-a', 'a', new LoroDoc())
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces')
-
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as {
-      workspaces: Array<{ workspaceId: string }>
-    }
-    expect(json.workspaces).toEqual([{ workspaceId: 'workspace-a' }])
-  })
-})
-
-describe('POST /api/workspaces/:workspaceId/canvases', () => {
-  beforeEach(() => {
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns { path } on success', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'new-canvas' }),
-    })
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as unknown
-    expect(json).toEqual({ path: 'new-canvas' })
-  })
-
-  it('returns 409 with Problem Details title on duplicate path', async () => {
-    const app = createCanvasRouter()
-    // Create once to seed the conflict.
-    await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'existing' }),
-    })
-    // Second creation must return Problem Details with a title field.
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'existing' }),
-    })
-    expect(res.status).toBe(409)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
-  })
-
-  it('returns 500 with Problem Details title when saveDocument fails unexpectedly', async () => {
-    // Block the canvas blob directory with a file so saveDocument throws a
-    // non-ConflictError, exercising the catch-all 500 branch (mutation-check
-    // guard for the 400 -> 500 change).
-    await mkdir(join(tmp.dir, 'blobs', 'ws1'), { recursive: true })
-    await writeFile(join(tmp.dir, 'blobs', 'ws1', 'canvas'), 'not-a-directory')
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'new-canvas' }),
-    })
-    expect(res.status).toBe(500)
-    const json = (await res.json()) as { title?: string }
-    expect(json.title).toBe('Failed to create canvas.')
-  })
-
-  it('returns 400 with Problem Details title on invalid path', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'bad path!' }),
-    })
-    expect(res.status).toBe(400)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-  })
-
-  it('returns 400 with Problem Details title when body is not JSON', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'text/plain' },
-      body: 'not-json',
-    })
-    expect(res.status).toBe(400)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
-  })
-
-  it('returns 400 with Problem Details title when body has no path field', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ name: 'oops' }),
-    })
-    expect(res.status).toBe(400)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
   })
 
   it('serves a nested document path at /api/w/:workspaceId/canvas/*', async () => {
@@ -263,2061 +140,5 @@ describe('POST /api/workspaces/:workspaceId/canvases', () => {
       method: 'DELETE',
     })
     expect(del.status).toBe(200)
-  })
-
-  it('creates a kind:markdown canvas, and the list carries it back', async () => {
-    const app = createCanvasRouter()
-    const createRes = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'note', kind: 'markdown' }),
-    })
-    expect(createRes.status).toBe(200)
-    // Response body stays exactly { path } — kind is not echoed back.
-    expect(await createRes.json()).toEqual({ path: 'note' })
-
-    const listRes = await app.request('/api/workspaces/ws1/canvases')
-    const listJson = (await listRes.json()) as { canvases: { path: string; kind: string }[] }
-    const created = listJson.canvases.find((c) => c.path === 'note')
-    expect(created?.kind).toBe('markdown')
-
-    // The empty LoroDoc a markdown-kind create saves loads without error —
-    // an empty doc is a valid initial document for either kind.
-    const snapshotRes = await app.request('/api/w/ws1/canvas/note/snapshot')
-    expect(snapshotRes.status).toBe(200)
-  })
-
-  it('creates a canvas without kind — response and list stay byte-identical to spatial back-compat', async () => {
-    const app = createCanvasRouter()
-    const createRes = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'legacy' }),
-    })
-    expect(createRes.status).toBe(200)
-    expect(await createRes.json()).toEqual({ path: 'legacy' })
-
-    const listRes = await app.request('/api/workspaces/ws1/canvases')
-    const listJson = (await listRes.json()) as { canvases: { path: string; kind: string }[] }
-    expect(listJson.canvases.find((c) => c.path === 'legacy')?.kind).toBe('spatial')
-  })
-
-  it('returns 400 with Problem Details title naming the actual failing field on an invalid kind', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'bad-kind', kind: 'bogus' }),
-    })
-    expect(res.status).toBe(400)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    // A valid path plus an invalid kind must not be told "path is required" —
-    // that names the wrong field and gives no path to recovery.
-    expect(json.title).toMatch(/kind/i)
-    expect(json.title).not.toMatch(/path is required/i)
-  })
-
-  it('falls back to the issue message when the invalid field is neither path nor kind', async () => {
-    // An array body parses as JSON but fails the object schema at the TOP level — path [] —
-    // exercising createCanvasRequestErrorTitle's fallback branch rather than a field-specific
-    // message that would name the wrong thing.
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/ws-a/canvases', {
-      method: 'POST',
-      headers: { 'content-type': 'application/json' },
-      body: JSON.stringify([1, 2, 3]),
-    })
-    expect(res.status).toBe(400)
-    const json = (await res.json()) as { title: string }
-    expect(json.title).not.toMatch(/path is required/i)
-    expect(json.title).not.toMatch(/spatial/i)
-    expect(json.title.length).toBeGreaterThan(0)
-  })
-
-  it('returns 400 with Problem Details { title } (not legacy { error, message }) on invalid workspaceId', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/bad.workspace/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'test' }),
-    })
-    expect(res.status).toBe(400)
-    const json = (await res.json()) as Record<string, unknown>
-    // Must have a Problem Details title, not the legacy { error, message } shape.
-    expect(typeof json.title).toBe('string')
-    expect(json.title as string).toBeTruthy()
-    // Must NOT carry the old shape keys.
-    expect(json).not.toHaveProperty('error')
-    expect(json).not.toHaveProperty('message')
-  })
-})
-
-describe('DELETE /api/workspaces/:workspaceId/canvases/:path', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns 200 { ok: true }, parses with deleteCanvasResponseSchema, and the canvas is gone from list/exists/snapshot', async () => {
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    const app = createCanvasRouter()
-
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a', {
-      method: 'DELETE',
-    })
-    expect(res.status).toBe(200)
-    const json: unknown = await res.json()
-    expect(deleteCanvasResponseSchema.parse(json)).toEqual({ ok: true })
-
-    const listRes = await app.request('/api/workspaces/session1/canvases')
-    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
-    expect(listJson.canvases.map((c) => c.path)).not.toContain('canvas-a')
-
-    const existsRes = await app.request('/api/w/session1/canvas/canvas-a/exists')
-    expect(await existsRes.json()).toEqual({ exists: false })
-
-    const snapshotRes = await app.request('/api/w/session1/canvas/canvas-a/snapshot')
-    expect(snapshotRes.status).toBe(404)
-  })
-
-  it('returns 404 with Problem Details { title } for a missing canvas', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/session1/canvases/never-created', {
-      method: 'DELETE',
-    })
-    expect(res.status).toBe(404)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
-    expect(json).not.toHaveProperty('error')
-  })
-
-  it('returns 400 with Problem Details { title } for an invalid workspaceId or path', async () => {
-    const app = createCanvasRouter()
-
-    const badWs = await app.request('/api/workspaces/bad.workspace/canvases/canvas-a', {
-      method: 'DELETE',
-    })
-    expect(badWs.status).toBe(400)
-    expect(typeof ((await badWs.json()) as { title?: string }).title).toBe('string')
-
-    const badPath = await app.request('/api/workspaces/session1/canvases/bad.path', {
-      method: 'DELETE',
-    })
-    expect(badPath.status).toBe(400)
-    expect(typeof ((await badPath.json()) as { title?: string }).title).toBe('string')
-  })
-
-  it('maps a thrown CorruptStoredDataError to 500 { error: corrupt_stored_data }, and only that error type — a plain throw stays a generic 500', async () => {
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    const canvasStore = await import('../store/document-store.js')
-    const spy = vi
-      .spyOn(canvasStore, 'deleteDocument')
-      .mockRejectedValueOnce(
-        corruptStoredData('/tmp/blobs/session1/canvas/abc.loro', 'broken canvas blob'),
-      )
-    const app = createCanvasRouter()
-    try {
-      const res = await app.request('/api/workspaces/session1/canvases/canvas-a', {
-        method: 'DELETE',
-      })
-      expect(res.status).toBe(500)
-      await expect(res.json()).resolves.toEqual({
-        error: 'corrupt_stored_data',
-        message: expect.stringContaining('broken canvas blob'),
-      })
-
-      // Mutation guard: a non-corruption throw must NOT hit the same
-      // branch — proves the mapping checks the error type, not "any throw".
-      spy.mockRejectedValueOnce(new Error('disk exploded'))
-      const res2 = await app.request('/api/workspaces/session1/canvases/canvas-a', {
-        method: 'DELETE',
-      })
-      expect(res2.status).toBe(500)
-      const json2: unknown = await res2.json()
-      expect(json2).not.toMatchObject({ error: 'corrupt_stored_data' })
-    } finally {
-      spy.mockRestore()
-    }
-  })
-
-  it('evicts the doc-cache on delete, so re-creating the same path after a warm cache read yields a fresh empty doc', async () => {
-    const app = createCanvasRouter()
-    await app.request('/api/workspaces/session1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'cached' }),
-    })
-    const clientDoc = new LoroDoc()
-    const prevVV = clientDoc.version()
-    const list = clientDoc.getMovableList('elements')
-    const m = list.insertContainer(0, new LoroMap())
-    m.set('id', 'old-element')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-    // Warm the doc-cache via the update path, matching real client traffic.
-    await app.request('/api/w/session1/canvas/cached/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-    expect(peekDoc('session1', 'cached')).toBeDefined()
-
-    const delRes = await app.request('/api/workspaces/session1/canvases/cached', {
-      method: 'DELETE',
-    })
-    expect(delRes.status).toBe(200)
-
-    // Re-creating must succeed (not 409) and must not resurrect the old doc.
-    const createRes = await app.request('/api/workspaces/session1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'cached' }),
-    })
-    expect(createRes.status).toBe(200)
-
-    const snapshotRes = await app.request('/api/w/session1/canvas/cached/snapshot')
-    expect(snapshotRes.status).toBe(200)
-    const buf = await snapshotRes.arrayBuffer()
-    const restored = LoroDoc.fromSnapshot(new Uint8Array(buf))
-    expect(restored.getMovableList('elements').length).toBe(0)
-  })
-})
-
-describe('PUT /api/workspaces/:workspaceId/canvases/:path/path', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns 200 { path }, the list shows the new path and not the old one, the old snapshot URL 404s, and re-creating the old path afterward succeeds as a fresh canvas', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    const app = createCanvasRouter()
-
-    const res = await app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'b' }),
-    })
-    expect(res.status).toBe(200)
-    const json: unknown = await res.json()
-    expect(renameDocumentPathResponseSchema.parse(json)).toEqual({ path: 'b' })
-
-    const listRes = await app.request('/api/workspaces/session1/canvases')
-    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
-    const paths = listJson.canvases.map((c) => c.path)
-    expect(paths).toContain('b')
-    expect(paths).not.toContain('a')
-
-    const oldSnapshotRes = await app.request('/api/w/session1/canvas/a/snapshot')
-    expect(oldSnapshotRes.status).toBe(404)
-
-    const recreateRes = await app.request('/api/workspaces/session1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'a' }),
-    })
-    expect(recreateRes.status).toBe(200)
-  })
-
-  it('returns 404 with Problem Details { title } for a missing source canvas', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/session1/canvases/never-created/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'target' }),
-    })
-    expect(res.status).toBe(404)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
-    expect(json).not.toHaveProperty('error')
-  })
-
-  it('returns 409 with Problem Details { title } when the target path is already taken', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'b', new LoroDoc())
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'b' }),
-    })
-    expect(res.status).toBe(409)
-    const json = (await res.json()) as { title?: string }
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
-
-    // Neither canvas was mutated by the rejected rename.
-    const listRes = await app.request('/api/workspaces/session1/canvases')
-    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
-    expect(listJson.canvases.map((c) => c.path).sort()).toEqual(['a', 'b'])
-  })
-
-  it('does not fork a phantom duplicate canvas when a rename races an in-flight /update that already resolved a doc reference through the old path', async () => {
-    const app = createCanvasRouter()
-    const baseDoc = new LoroDoc()
-    baseDoc.getText('content').insert(0, 'original')
-    baseDoc.commit()
-    await saveDocument('session1', 'a', baseDoc)
-
-    // A client update built against the pre-rename base.
-    const clientDoc = LoroDoc.fromSnapshot(baseDoc.export({ mode: 'snapshot' }))
-    const fromVV = clientDoc.version()
-    clientDoc.getText('content').insert('original'.length, ' + edit')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: fromVV })
-
-    // Stall the /update route's getDoc() call so a rename can be fired
-    // while the request is paused mid-flight, matching the real race: the
-    // read resolves before the rename runs, the write happens after.
-    let releaseGetDoc: () => void = () => undefined
-    const getDocGate = new Promise<void>((resolve) => {
-      releaseGetDoc = resolve
-    })
-    let signalGetDocCalled: () => void = () => undefined
-    const getDocCalled = new Promise<void>((resolve) => {
-      signalGetDocCalled = resolve
-    })
-    const actual =
-      await vi.importActual<typeof import('../store/doc-cache.js')>('../store/doc-cache.js')
-    vi.mocked(getDoc).mockImplementationOnce(async (workspaceId, path) => {
-      signalGetDocCalled()
-      await getDocGate
-      return actual.getDoc(workspaceId, path)
-    })
-
-    const updatePromise = app.request('/api/w/session1/canvas/a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-
-    await getDocCalled
-
-    // Fire the rename while the update is stalled mid-flight.
-    const renamePromise = app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'b' }),
-    })
-    // Give the rename a chance to run before letting the stalled read continue.
-    await new Promise((r) => setTimeout(r, 20))
-    releaseGetDoc()
-
-    const [updateRes, renameRes] = await Promise.all([updatePromise, renamePromise])
-    expect(renameRes.status).toBe(200)
-    expect(updateRes.status).toBe(200)
-
-    // Exactly one canvas must survive -- the update must not have silently
-    // inserted a phantom duplicate back at the old path.
-    const listRes = await app.request('/api/workspaces/session1/canvases')
-    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
-    expect(listJson.canvases.map((c) => c.path)).toEqual(['b'])
-  })
-
-  it('returns 400 with Problem Details { title } for invalid input shapes', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    const app = createCanvasRouter()
-
-    const noBody = await app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-    })
-    expect(noBody.status).toBe(400)
-
-    const badNewPath = await app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'bad.path' }),
-    })
-    expect(badNewPath.status).toBe(400)
-
-    const emptyPath = await app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: '' }),
-    })
-    expect(emptyPath.status).toBe(400)
-
-    const badWorkspace = await app.request('/api/workspaces/bad.workspace/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'b' }),
-    })
-    expect(badWorkspace.status).toBe(400)
-
-    const badCurrentPath = await app.request('/api/workspaces/session1/canvases/bad.path/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'b' }),
-    })
-    expect(badCurrentPath.status).toBe(400)
-
-    for (const res of [noBody, badNewPath, emptyPath, badWorkspace, badCurrentPath]) {
-      const body = (await res.json()) as { title?: string }
-      expect(typeof body.title).toBe('string')
-      expect(body.title!.length).toBeGreaterThan(0)
-    }
-  })
-
-  it('evicts the doc-cache on rename: updating via the OLD path afterward lazily creates a FRESH canvas rather than resurrecting the warmed doc, while the new path keeps the real content', async () => {
-    const app = createCanvasRouter()
-    await app.request('/api/workspaces/session1/canvases', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'a' }),
-    })
-    const clientDoc = new LoroDoc()
-    const prevVV = clientDoc.version()
-    const list = clientDoc.getMovableList('elements')
-    const m = list.insertContainer(0, new LoroMap())
-    m.set('id', 'old-element')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-    // Warm the doc-cache via the update path, matching real client traffic.
-    await app.request('/api/w/session1/canvas/a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-    expect(peekDoc('session1', 'a')).toBeDefined()
-
-    const renameRes = await app.request('/api/workspaces/session1/canvases/a/path', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ path: 'b' }),
-    })
-    expect(renameRes.status).toBe(200)
-
-    // The renamed canvas keeps its content, now reachable at the new path.
-    const bSnapshotRes = await app.request('/api/w/session1/canvas/b/snapshot')
-    expect(bSnapshotRes.status).toBe(200)
-    const bBuf = await bSnapshotRes.arrayBuffer()
-    const bRestored = LoroDoc.fromSnapshot(new Uint8Array(bBuf))
-    expect(
-      (bRestored.getMovableList('elements').toJSON() as { id: string }[]).map((e) => e.id),
-    ).toEqual(['old-element'])
-
-    // Updating through the OLD path lazily creates a FRESH canvas — it must
-    // not resurrect the evicted doc's history.
-    const anotherClientDoc = new LoroDoc()
-    const anotherPrevVV = anotherClientDoc.version()
-    const anotherList = anotherClientDoc.getMovableList('elements')
-    const anotherM = anotherList.insertContainer(0, new LoroMap())
-    anotherM.set('id', 'fresh-element')
-    anotherClientDoc.commit()
-    const anotherUpdate = anotherClientDoc.export({ mode: 'update', from: anotherPrevVV })
-    await app.request('/api/w/session1/canvas/a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: anotherUpdate,
-    })
-
-    const freshSnapshotRes = await app.request('/api/w/session1/canvas/a/snapshot')
-    expect(freshSnapshotRes.status).toBe(200)
-    const freshBuf = await freshSnapshotRes.arrayBuffer()
-    const freshRestored = LoroDoc.fromSnapshot(new Uint8Array(freshBuf))
-    const freshIds = (freshRestored.getMovableList('elements').toJSON() as { id: string }[]).map(
-      (e) => e.id,
-    )
-    expect(freshIds).toEqual(['fresh-element'])
-    expect(freshIds).not.toContain('old-element')
-  })
-
-  it('maps a thrown CorruptStoredDataError to 500 { error: corrupt_stored_data }, and only that error type — a plain throw stays a generic 500', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    const canvasStore = await import('../store/document-store.js')
-    const spy = vi
-      .spyOn(canvasStore, 'renameDocumentPath')
-      .mockRejectedValueOnce(
-        corruptStoredData('/tmp/blobs/session1/canvas/abc.loro', 'broken canvas blob'),
-      )
-    const app = createCanvasRouter()
-    try {
-      const res = await app.request('/api/workspaces/session1/canvases/a/path', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ path: 'b' }),
-      })
-      expect(res.status).toBe(500)
-      await expect(res.json()).resolves.toEqual({
-        error: 'corrupt_stored_data',
-        message: expect.stringContaining('broken canvas blob'),
-      })
-
-      // Mutation guard: a non-corruption throw must NOT hit the same
-      // branch — proves the mapping checks the error type, not "any throw".
-      spy.mockRejectedValueOnce(new Error('disk exploded'))
-      const res2 = await app.request('/api/workspaces/session1/canvases/a/path', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ path: 'b' }),
-      })
-      expect(res2.status).toBe(500)
-      const json2: unknown = await res2.json()
-      expect(json2).not.toMatchObject({ error: 'corrupt_stored_data' })
-    } finally {
-      spy.mockRestore()
-    }
-  })
-})
-
-describe('GET /api/workspaces/:workspaceId/canvases', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns the canvas list', async () => {
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    await saveDocument('session1', 'canvas-b', new LoroDoc())
-
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/session1/canvases')
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as { canvases: { path: string }[] }
-    const paths = json.canvases.map((c) => c.path)
-    expect(paths).toContain('canvas-a')
-    expect(paths).toContain('canvas-b')
-  })
-
-  it('returns 400 for an invalid workspaceId', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/bad.sid/canvases')
-    expect(res.status).toBe(400)
-  })
-
-  it('returns 404 for a workspace that does not exist, not an empty list', async () => {
-    // A 200-empty here is a false statement of fact: the caller cannot tell
-    // "this workspace has no canvases" from "you asked about a workspace this
-    // daemon has never heard of", and the web app renders the first reading —
-    // an empty state with a Create button — for what is actually the second.
-    // A stale pairing (a workspace id minted by an earlier install and kept
-    // in the browser's localStorage) then looks exactly like data loss.
-    // The v1 document routes already 404 an unknown workspace; this brings
-    // the legacy list route into agreement.
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/never-registered/canvases')
-    expect(res.status).toBe(404)
-  })
-
-  // listDocuments no longer walks per-workspace directories, so the previous
-  // "broken session directory" 500 case no longer applies.
-})
-
-describe('GET /api/workspaces/:workspaceId/canvases', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'workspace1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('also returns the canvas list from the canonical workspace route', async () => {
-    await saveDocument('workspace1', 'canvas-a', new LoroDoc())
-
-    const app = createCanvasRouter()
-    const res = await app.request('/api/workspaces/workspace1/canvases')
-
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as { canvases: { path: string }[] }
-    expect(json.canvases).toEqual([expect.objectContaining({ path: 'canvas-a' })])
-  })
-})
-
-describe('GET /api/w/:workspaceId/canvas/:path/snapshot', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns the Loro snapshot binary', async () => {
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    const map = list.insertContainer(0, new LoroMap())
-    map.set('id', 'elem-001')
-    doc.commit()
-    await saveDocument('session1', 'canvas-a', doc)
-
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/canvas-a/snapshot')
-    expect(res.status).toBe(200)
-    expect(res.headers.get('Content-Type')).toContain('application/octet-stream')
-
-    const buf = await res.arrayBuffer()
-    expect(buf.byteLength).toBeGreaterThan(0)
-
-    // Restore the binary and confirm the element is present.
-    const restored = LoroDoc.fromSnapshot(new Uint8Array(buf))
-    const elements = restored.getMovableList('elements').toJSON() as { id: string }[]
-    expect(elements).toHaveLength(1)
-    expect(elements[0].id).toBe('elem-001')
-  })
-
-  it('returns 400 for an invalid path', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/bad.path/snapshot')
-    expect(res.status).toBe(400)
-  })
-
-  it('returns 404 with Problem Details { title } for a canvas that was never created', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/never-created/snapshot')
-    expect(res.status).toBe(404)
-    const json = (await res.json()) as { title?: string }
-    // Same shape as DELETE's 404 — the client parses problem-details for
-    // both routes — deliberately not thumbnails/restore's { error, message }.
-    expect(typeof json.title).toBe('string')
-    expect(json.title!.length).toBeGreaterThan(0)
-    expect(json).not.toHaveProperty('error')
-  })
-
-  it('returns 404 for a canvas that existed and was deleted', async () => {
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    const app = createCanvasRouter()
-    await app.request('/api/workspaces/session1/canvases/canvas-a', { method: 'DELETE' })
-
-    const res = await app.request('/api/w/session1/canvas/canvas-a/snapshot')
-    expect(res.status).toBe(404)
-  })
-})
-
-describe('GET /api/w/:workspaceId/canvas/:path/exists', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns exists:true for a canvas that was actually saved', async () => {
-    const doc = new LoroDoc()
-    await saveDocument('session1', 'canvas-a', doc)
-
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/canvas-a/exists')
-    expect(res.status).toBe(200)
-    expect(await res.json()).toEqual({ exists: true })
-  })
-
-  it('returns exists:false for an unregistered canvas without creating it', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/never-created/exists')
-    expect(res.status).toBe(200)
-    expect(await res.json()).toEqual({ exists: false })
-
-    // GET /exists must never have the getDoc/loadDocument side effect that
-    // /snapshot has: the canvas should still be absent afterward.
-    const followUp = await app.request('/api/w/session1/canvas/never-created/exists')
-    expect(await followUp.json()).toEqual({ exists: false })
-  })
-
-  it('returns 400 for an invalid path', async () => {
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/bad.path/exists')
-    expect(res.status).toBe(400)
-  })
-})
-
-describe('POST /api/workspaces/:workspaceId/canvases/:path/compact', () => {
-  function createVersionStoreMock() {
-    return {
-      save: vi.fn(),
-      load: vi.fn(),
-      list: vi.fn(),
-      saveThumbnail: vi.fn(),
-      loadThumbnail: vi.fn(),
-      earliestFrontiers: vi.fn().mockResolvedValue([]),
-      getFrontiersBase64: vi.fn(),
-      renameBranchInVersions: vi.fn(),
-      pruneSandwichedAutoVersions: vi.fn().mockResolvedValue({ deletedCount: 0, deletedIds: [] }),
-    }
-  }
-
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns structured 500 for a broken snapshot', async () => {
-    const { getDb } = await import('../store/db/index.js')
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    const db = await getDb(tmp.dir)
-    const row = await db
-      .selectFrom('documents')
-      .select(['id'])
-      .where('workspaceId', '=', 'session1')
-      .where('path', '=', 'canvas-a')
-      .executeTakeFirstOrThrow()
-    const blobPath = join(tmp.dir, 'blobs', 'session1', 'canvas', `${row.id}.loro`)
-    await writeFile(blobPath, Buffer.from('not-a-loro-snapshot'))
-
-    const app = createCanvasRouter({ versionStore: createVersionStoreMock() })
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a/compact', {
-      method: 'POST',
-    })
-
-    expect(res.status).toBe(500)
-    await expect(res.json()).resolves.toEqual({
-      error: 'corrupt_stored_data',
-      message: expect.stringContaining(`${row.id}.loro`),
-    })
-  })
-
-  // Canvas blobs live under blobs/{workspaceId}/canvas/, so the previous
-  // "non-directory session path" stat failure case no longer maps. compact
-  // returns no-file for missing blobs, and the corrupt-snapshot case above
-  // still exercises the corruption branch.
-})
-
-describe('POST /api/workspaces/:workspaceId/canvases/optimize-all', () => {
-  function createVersionStoreMock() {
-    return {
-      save: vi.fn(),
-      load: vi.fn(),
-      list: vi.fn(),
-      saveThumbnail: vi.fn(),
-      loadThumbnail: vi.fn(),
-      // No version cut available — each per-canvas compact returns
-      // reason: 'no-versions'. That is the realistic dry-run shape; what
-      // matters is the bulk endpoint loops every canvas and aggregates
-      // totals correctly.
-      earliestFrontiers: vi.fn().mockResolvedValue(null),
-      getFrontiersBase64: vi.fn(),
-      renameBranchInVersions: vi.fn(),
-      // The bulk-optimize route doesn't exercise prune by default, but the
-      // VersionStore contract requires it. Returning a no-op result keeps
-      // the mock fully type-compatible so a future route change that does
-      // call this method here cannot trip on a missing key.
-      pruneSandwichedAutoVersions: vi.fn().mockResolvedValue({ deletedCount: 0, deletedIds: [] }),
-    }
-  }
-
-  beforeEach(() => {
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('iterates every canvas in the workspace and returns aggregated results', async () => {
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    await saveDocument('session1', 'canvas-b', new LoroDoc())
-
-    const app = createCanvasRouter({ versionStore: createVersionStoreMock() })
-    const res = await app.request('/api/workspaces/session1/canvases/optimize-all', {
-      method: 'POST',
-    })
-
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as {
-      results: Array<{
-        path: string
-        compacted: boolean
-        beforeBytes: number
-        afterBytes: number
-        reason?: string
-      }>
-      totalBeforeBytes: number
-      totalAfterBytes: number
-    }
-    expect(json.results.map((r) => r.path).sort()).toEqual(['canvas-a', 'canvas-b'])
-    expect(json.results.every((r) => r.compacted === false)).toBe(true)
-    expect(json.results.every((r) => r.reason === 'no-versions')).toBe(true)
-    expect(json.totalBeforeBytes).toBeGreaterThan(0)
-    expect(json.totalAfterBytes).toBe(json.totalBeforeBytes)
-  })
-
-  it('returns an empty result set when the workspace has no canvases', async () => {
-    const app = createCanvasRouter({ versionStore: createVersionStoreMock() })
-    const res = await app.request('/api/workspaces/session1/canvases/optimize-all', {
-      method: 'POST',
-    })
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as {
-      results: unknown[]
-      totalBeforeBytes: number
-      totalAfterBytes: number
-    }
-    expect(json.results).toEqual([])
-    expect(json.totalBeforeBytes).toBe(0)
-    expect(json.totalAfterBytes).toBe(0)
-  })
-
-  it('prune-sandwiched delegates to versionStore.pruneSandwichedAutoVersions for every canvas', async () => {
-    await saveDocument('session1', 'canvas-a', new LoroDoc())
-    await saveDocument('session1', 'canvas-b', new LoroDoc())
-
-    const versionStore = createVersionStoreMock()
-    versionStore.pruneSandwichedAutoVersions = vi
-      .fn()
-      .mockImplementation(async (_wid: string, path: string) => ({
-        deletedCount: path === 'canvas-a' ? 2 : 1,
-        deletedIds: path === 'canvas-a' ? ['x', 'y'] : ['z'],
-      }))
-
-    const app = createCanvasRouter({ versionStore })
-    const res = await app.request('/api/workspaces/session1/versions/prune-sandwiched', {
-      method: 'POST',
-    })
-
-    expect(res.status).toBe(200)
-    const json = (await res.json()) as {
-      results: Array<{ path: string; deletedCount: number }>
-      totalDeleted: number
-    }
-    expect(json.totalDeleted).toBe(3)
-    expect(json.results.map((r) => r.path).sort()).toEqual(['canvas-a', 'canvas-b'])
-    expect(versionStore.pruneSandwichedAutoVersions).toHaveBeenCalledTimes(2)
-  })
-})
-
-// names-store now lives in the sqlite metadata DB; the corruption-on-disk
-// failure modes covered above no longer apply. DB-side error handling is
-// exercised through unit tests on names-store directly.
-
-describe('POST /api/w/:workspaceId/canvas/:path/update', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('applies a Loro update to the document', async () => {
-    // Create the change in the client-side doc.
-    const clientDoc = new LoroDoc()
-    const prevVV = clientDoc.version()
-    const list = clientDoc.getMovableList('elements')
-    const map = list.insertContainer(0, new LoroMap())
-    map.set('id', 'elem-from-client')
-    map.set('type', 'ellipse')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-
-    const app = createCanvasRouter()
-    const res = await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-    expect(res.status).toBe(200)
-    // Pin the shared contract: the response body must parse under the
-    // Zod schema the web import client consumes, not just an ad-hoc shape.
-    const body: unknown = await res.json()
-    expect(updateCanvasResponseSchema.safeParse(body).success).toBe(true)
-
-    // Clear the cache, reload, and confirm the change was persisted.
-    clearCache()
-    const { loadDocument } = await import('../store/document-store.js')
-    const serverDoc = await loadDocument('session1', 'canvas-a')
-    const elements = serverDoc.getMovableList('elements').toJSON() as { id: string; type: string }[]
-    expect(elements).toHaveLength(1)
-    expect(elements[0].id).toBe('elem-from-client')
-    expect(elements[0].type).toBe('ellipse')
-  })
-
-  it('rejects an oversized update body with 413 payload_too_large', async () => {
-    const app = createCanvasRouter()
-    const oversized = new Uint8Array(16 * 1024 * 1024 + 1)
-    const res = await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: oversized,
-    })
-    expect(res.status).toBe(413)
-    const body: unknown = await res.json()
-    expect(body).toMatchObject({ error: 'payload_too_large' })
-  })
-
-  it('evicts the cached doc when saveDocument fails, so a subsequent read does not serve the unpersisted update', async () => {
-    const app = createCanvasRouter()
-
-    // Persist an initial one-element state directly.
-    const initial = new LoroDoc()
-    const initialList = initial.getMovableList('elements')
-    const initialElem = initialList.insertContainer(0, new LoroMap())
-    initialElem.set('id', 'persisted')
-    initial.commit()
-    await saveDocument('session1', 'canvas-a', initial)
-
-    // Pull it into the doc-cache so there is a live cached doc to poison.
-    await app.request('/api/w/session1/canvas/canvas-a/snapshot')
-    expect(peekDoc('session1', 'canvas-a')).toBeDefined()
-
-    // Build a client update that adds a second element.
-    const clientDoc = LoroDoc.fromSnapshot(initial.export({ mode: 'snapshot' }))
-    const prevVV = clientDoc.version()
-    const clientList = clientDoc.getMovableList('elements')
-    const clientElem = clientList.insertContainer(clientList.length, new LoroMap())
-    clientElem.set('id', 'unpersisted')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-
-    // Force the persistence step inside saveDocument to fail after doc.import()
-    // has already mutated the cached doc.
-    const exportSpy = vi.spyOn(LoroDoc.prototype, 'export').mockImplementationOnce(() => {
-      throw new Error('simulated snapshot failure')
-    })
-    const res = await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-    exportSpy.mockRestore()
-
-    expect(res.status).toBe(500)
-    // The cache must be evicted: without eviction, the in-memory doc already
-    // absorbed the update even though saveDocument never persisted it.
-    expect(peekDoc('session1', 'canvas-a')).toBeUndefined()
-
-    const reloaded = await getDoc('session1', 'canvas-a')
-    const ids = (reloaded.getMovableList('elements').toJSON() as Array<{ id: string }>).map(
-      (el) => el.id,
-    )
-    expect(ids).toEqual(['persisted'])
-  })
-})
-
-// Version API coverage: auto-save on update, list, manual save, and restore.
-describe('versions API', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('saves an auto-version immediately when autoVersionIntervalMs=0', async () => {
-    const clientDoc = new LoroDoc()
-    const prevVV = clientDoc.version()
-    const list = clientDoc.getMovableList('elements')
-    const m = list.insertContainer(0, new LoroMap())
-    m.set('id', 'e1')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-
-    const app = createCanvasRouter({ autoVersionIntervalMs: 0 })
-    const resUpdate = await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-    expect(resUpdate.status).toBe(200)
-
-    // Auto-version saving is best-effort and async, so wait briefly.
-    await new Promise((r) => setTimeout(r, 50))
-
-    const resList = await app.request('/api/workspaces/session1/canvases/canvas-a/versions')
-    expect(resList.status).toBe(200)
-    const body = (await resList.json()) as {
-      versions: Array<{ auto: boolean; elementCount: number }>
-    }
-    expect(body.versions.length).toBeGreaterThanOrEqual(1)
-    expect(body.versions[0].auto).toBe(true)
-    expect(body.versions[0].elementCount).toBe(1)
-  })
-
-  it('saves a manual version with a label through POST /versions', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'before refactor' }),
-    })
-    expect(res.status).toBe(200)
-    const body = (await res.json()) as { version: { auto: boolean; label?: string } }
-    expect(body.version.auto).toBe(false)
-    expect(body.version.label).toBe('before refactor')
-  })
-
-  it('POST /versions persists an explicit operator', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        label: 'ai save',
-        operator: {
-          kind: 'ai',
-          peerId: 'peer-ai',
-          displayName: 'Assistant',
-          agentId: 'agent-1',
-        },
-      }),
-    })
-
-    expect(res.status).toBe(200)
-    const body = (await res.json()) as {
-      version: {
-        operator?: { kind: string; peerId: string; displayName?: string; agentId?: string }
-      }
-    }
-    expect(body.version.operator).toEqual({
-      kind: 'ai',
-      peerId: 'peer-ai',
-      displayName: 'Assistant',
-      agentId: 'agent-1',
-    })
-
-    const listRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions')
-    const listBody = (await listRes.json()) as {
-      versions: Array<{
-        operator?: { kind: string; peerId: string; displayName?: string; agentId?: string }
-      }>
-    }
-    expect(listBody.versions[0]?.operator).toEqual(body.version.operator)
-  })
-
-  it('POST /versions defaults operator to human when omitted', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'manual save' }),
-    })
-
-    expect(res.status).toBe(200)
-    const body = (await res.json()) as {
-      version: { operator?: { kind: string; peerId: string; displayName?: string } }
-    }
-    expect(body.version.operator?.kind).toBe('human')
-    expect(body.version.operator?.peerId).toMatch(/\S+/)
-    expect(body.version.operator?.displayName).toBe(userInfo().username)
-  })
-
-  it('POST /update auto-version persists system/auto-save operator', async () => {
-    const clientDoc = new LoroDoc()
-    const prevVV = clientDoc.version()
-    const list = clientDoc.getMovableList('elements')
-    const m = list.insertContainer(0, new LoroMap())
-    m.set('id', 'e-auto')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-
-    const app = createCanvasRouter({ autoVersionIntervalMs: 0 })
-    const resUpdate = await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-    expect(resUpdate.status).toBe(200)
-
-    await new Promise((r) => setTimeout(r, 50))
-
-    const resList = await app.request('/api/workspaces/session1/canvases/canvas-a/versions')
-    const body = (await resList.json()) as {
-      versions: Array<{ operator?: { kind: string; peerId: string; displayName?: string } }>
-    }
-    expect(body.versions[0]?.operator).toMatchObject({
-      kind: 'system',
-      displayName: 'auto-save',
-    })
-    expect(body.versions[0]?.operator?.peerId).toMatch(/\S+/)
-  })
-
-  it('restores the past-state element set through POST /versions/:id/restore', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-    // Step 1: Write the initial one-element state through /update.
-    const initial = new LoroDoc()
-    const vv0 = initial.version()
-    const list = initial.getMovableList('elements')
-    const m0 = list.insertContainer(list.length, new LoroMap())
-    m0.set('id', 'keep-me')
-    m0.set('type', 'rectangle')
-    initial.commit()
-    await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: initial.export({ mode: 'update', from: vv0 }),
-    })
-
-    // Step 2: Save the current one-element state as v1.
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-    const v1id = saveBody.version.id
-
-    // Step 3: Add two more elements after v1.
-    const vv1 = initial.version()
-    const m1 = list.insertContainer(list.length, new LoroMap())
-    m1.set('id', 'added-1')
-    m1.set('type', 'ellipse')
-    const m2 = list.insertContainer(list.length, new LoroMap())
-    m2.set('id', 'added-2')
-    m2.set('type', 'diamond')
-    initial.commit()
-    await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: initial.export({ mode: 'update', from: vv1 }),
-    })
-
-    // Step 4: Restore v1. The CRDT merge imports the past snapshot.
-    const resRestore = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${v1id}/restore`,
-      { method: 'POST' },
-    )
-    expect(resRestore.status).toBe(200)
-  })
-
-  it('evicts the cached doc when saveDocument fails during in-place restore, so a subsequent read does not serve the un-persisted reconcile', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-    // Step 1: seed with keep-me.
-    const initial = new LoroDoc()
-    const vv0 = initial.version()
-    const list = initial.getMovableList('elements')
-    const m0 = list.insertContainer(0, new LoroMap())
-    m0.set('id', 'keep-me')
-    initial.commit()
-    await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: initial.export({ mode: 'update', from: vv0 }),
-    })
-
-    // Step 2: save v1.
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-    // Step 3: add a second element after v1, so restoring v1 would tombstone it.
-    const vv1 = initial.version()
-    const m1 = list.insertContainer(list.length, new LoroMap())
-    m1.set('id', 'added-after-v1')
-    initial.commit()
-    await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: initial.export({ mode: 'update', from: vv1 }),
-    })
-
-    expect(peekDoc('session1', 'canvas-a')).toBeDefined()
-
-    // Restore calls doc.export() twice before the failure point we care about:
-    // once inside versionStore.load() to clone the live doc, then again inside
-    // saveDocument() once reconcile+commit have already mutated the cached doc.
-    // Let the first pass through and fail only the second.
-    let exportCallCount = 0
-    const originalExport = LoroDoc.prototype.export
-    const exportSpy = vi.spyOn(LoroDoc.prototype, 'export').mockImplementation(function (
-      this: LoroDoc,
-      ...args: Parameters<typeof originalExport>
-    ) {
-      exportCallCount++
-      if (exportCallCount === 2) {
-        throw new Error('simulated snapshot failure')
-      }
-      return originalExport.apply(this, args)
-    })
-
-    const restoreRes = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-      { method: 'POST' },
-    )
-    exportSpy.mockRestore()
-
-    expect(restoreRes.status).toBe(500)
-    // The cache must be evicted: without eviction, the reconciled-but-never-
-    // saved tombstone state would keep serving from memory.
-    expect(peekDoc('session1', 'canvas-a')).toBeUndefined()
-
-    const reloaded = await getDoc('session1', 'canvas-a')
-    const ids = (reloaded.getMovableList('elements').toJSON() as Array<{ id: string }>)
-      .map((el) => el.id)
-      .sort()
-    expect(ids).toEqual(['added-after-v1', 'keep-me'])
-  })
-
-  it('evicts the cached doc when reconcile/commit itself fails during in-place restore, not only when saveDocument fails', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-    const initial = new LoroDoc()
-    const vv0 = initial.version()
-    const list = initial.getMovableList('elements')
-    const m0 = list.insertContainer(0, new LoroMap())
-    m0.set('id', 'keep-me')
-    initial.commit()
-    await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: initial.export({ mode: 'update', from: vv0 }),
-    })
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-    const vv1 = initial.version()
-    const m1 = list.insertContainer(list.length, new LoroMap())
-    m1.set('id', 'added-after-v1')
-    initial.commit()
-    await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: initial.export({ mode: 'update', from: vv1 }),
-    })
-
-    expect(peekDoc('session1', 'canvas-a')).toBeDefined()
-
-    // doc.import already mutated the live cached doc by the time
-    // doc.commit() runs, so a throw here must still evict the cache -- not
-    // only the saveDocument failure path further down.
-    const commitSpy = vi.spyOn(LoroDoc.prototype, 'commit').mockImplementationOnce(() => {
-      throw new Error('simulated commit failure')
-    })
-
-    const restoreRes = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-      { method: 'POST' },
-    )
-    commitSpy.mockRestore()
-
-    expect(restoreRes.status).toBe(500)
-    expect(peekDoc('session1', 'canvas-a')).toBeUndefined()
-
-    const reloaded = await getDoc('session1', 'canvas-a')
-    const ids = (reloaded.getMovableList('elements').toJSON() as Array<{ id: string }>)
-      .map((el) => el.id)
-      .sort()
-    expect(ids).toEqual(['added-after-v1', 'keep-me'])
-  })
-
-  it('returns 404 when restoring a missing version id', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request(
-      '/api/workspaces/session1/canvases/canvas-a/versions/nonexistent/restore',
-      { method: 'POST' },
-    )
-    expect(res.status).toBe(404)
-  })
-
-  it('returns 400 for an invalid version id', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request(
-      '/api/workspaces/session1/canvases/canvas-a/versions/bad.id/restore',
-      { method: 'POST' },
-    )
-    expect(res.status).toBe(400)
-  })
-
-  // targetPath + overwrite must reconcile onto the target's LIVE cached doc
-  // instead of replacing persistence, so connected clients stay on the same
-  // CRDT lineage and converge through the normal update broadcast.
-  describe('overwrite restore reconciles instead of replacing', () => {
-    afterEach(() => {
-      setBroadcastFn(() => {})
-    })
-
-    it('targetPath === path with overwrite:true reconciles the live doc in place and broadcasts an update', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      // v1: one element.
-      const initial = new LoroDoc()
-      const vv0 = initial.version()
-      const list = initial.getMovableList('elements')
-      const m0 = list.insertContainer(0, new LoroMap())
-      m0.set('id', 'keep-me')
-      initial.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: initial.export({ mode: 'update', from: vv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      // Add a second element after v1.
-      const vv1 = initial.version()
-      const m1 = list.insertContainer(list.length, new LoroMap())
-      m1.set('id', 'added-after-v1')
-      initial.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: initial.export({ mode: 'update', from: vv1 }),
-      })
-
-      const docBefore = peekDoc('session1', 'canvas-a')
-      expect(docBefore).toBeDefined()
-
-      const broadcastCalls: Array<{ workspaceId: string; path: string; byteLength: number }> = []
-      setBroadcastFn((workspaceId, path, update) => {
-        broadcastCalls.push({ workspaceId, path, byteLength: update.byteLength })
-      })
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-a', overwrite: true }),
-        },
-      )
-      expect(restoreRes.status).toBe(200)
-
-      // Same object identity: the cached doc was mutated in place, not
-      // replaced by a differently-lineaged document loaded from `past`.
-      expect(peekDoc('session1', 'canvas-a')).toBe(docBefore)
-
-      expect(broadcastCalls).toHaveLength(1)
-      expect(broadcastCalls[0]?.workspaceId).toBe('session1')
-      expect(broadcastCalls[0]?.path).toBe('canvas-a')
-      expect(broadcastCalls[0]?.byteLength).toBeGreaterThan(0)
-    })
-
-    it('targetPath === path WITHOUT overwrite still restores in place (same-path is never treated as a distinct target)', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      const initial = new LoroDoc()
-      const vv0 = initial.version()
-      const list = initial.getMovableList('elements')
-      const m0 = list.insertContainer(0, new LoroMap())
-      m0.set('id', 'keep-me')
-      initial.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: initial.export({ mode: 'update', from: vv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-a' }),
-        },
-      )
-      expect(restoreRes.status).toBe(200)
-      const body = (await restoreRes.json()) as { ok: boolean }
-      expect(body.ok).toBe(true)
-    })
-
-    it('restoring into a different existing canvas reconciles that canvas and broadcasts to it, not the source', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      // Source canvas-a: v1 has only "keep-me".
-      const sourceDoc = new LoroDoc()
-      const svv0 = sourceDoc.version()
-      const sourceList = sourceDoc.getMovableList('elements')
-      const sm0 = sourceList.insertContainer(0, new LoroMap())
-      sm0.set('id', 'keep-me')
-      sourceDoc.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: sourceDoc.export({ mode: 'update', from: svv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      // Target canvas-b already exists with an unrelated element.
-      const targetDoc = new LoroDoc()
-      const tvv0 = targetDoc.version()
-      const targetList = targetDoc.getMovableList('elements')
-      const tm0 = targetList.insertContainer(0, new LoroMap())
-      tm0.set('id', 'b-only')
-      targetDoc.commit()
-      await app.request('/api/w/session1/canvas/canvas-b/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: targetDoc.export({ mode: 'update', from: tvv0 }),
-      })
-
-      const broadcastCalls: Array<{ path: string }> = []
-      setBroadcastFn((_workspaceId, path) => {
-        broadcastCalls.push({ path })
-      })
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-b', overwrite: true }),
-        },
-      )
-      expect(restoreRes.status).toBe(200)
-      const restoreBody = (await restoreRes.json()) as { documentId: string; elementCount: number }
-      expect(restoreBody.documentId).toBe('session1/canvas-b')
-
-      expect(broadcastCalls).toHaveLength(1)
-      expect(broadcastCalls[0]?.path).toBe('canvas-b')
-    })
-
-    it('restoring into a new (non-existent) target path still creates it', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      const sourceDoc = new LoroDoc()
-      const svv0 = sourceDoc.version()
-      const sourceList = sourceDoc.getMovableList('elements')
-      const sm0 = sourceList.insertContainer(0, new LoroMap())
-      sm0.set('id', 'keep-me')
-      sourceDoc.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: sourceDoc.export({ mode: 'update', from: svv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-new' }),
-        },
-      )
-      expect(restoreRes.status).toBe(200)
-      const restoreBody = (await restoreRes.json()) as { documentId: string; elementCount: number }
-      expect(restoreBody.documentId).toBe('session1/canvas-new')
-      // The restored doc has one alive legacy element ("keep-me"); the
-      // fixed countAliveNodes reports the real count instead of the
-      // retired countElements(_doc) stub's always-0.
-      expect(restoreBody.elementCount).toBe(1)
-    })
-
-    it('restoring a markdown-kind canvas into a new target path carries the source kind forward, not the spatial default', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      await app.request('/api/workspaces/session1/canvases', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ path: 'canvas-a', kind: 'markdown' }),
-      })
-
-      const sourceDoc = new LoroDoc()
-      const svv0 = sourceDoc.version()
-      sourceDoc.getText('body').insert(0, 'hello')
-      sourceDoc.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: sourceDoc.export({ mode: 'update', from: svv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-new' }),
-        },
-      )
-      expect(restoreRes.status).toBe(200)
-
-      const listRes = await app.request('/api/workspaces/session1/canvases')
-      const listBody = (await listRes.json()) as {
-        canvases: { path: string; kind: string }[]
-      }
-      expect(listBody.canvases.find((c) => c.path === 'canvas-new')?.kind).toBe('markdown')
-    })
-
-    it('restoring a markdown-kind canvas onto an existing spatial-kind target syncs the target kind to match the restored content', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      await app.request('/api/workspaces/session1/canvases', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ path: 'canvas-a', kind: 'markdown' }),
-      })
-      await app.request('/api/workspaces/session1/canvases', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ path: 'canvas-b', kind: 'spatial' }),
-      })
-
-      const sourceDoc = new LoroDoc()
-      const svv0 = sourceDoc.version()
-      sourceDoc.getText('body').insert(0, 'hello')
-      sourceDoc.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: sourceDoc.export({ mode: 'update', from: svv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-b', overwrite: true }),
-        },
-      )
-      expect(restoreRes.status).toBe(200)
-
-      const listRes = await app.request('/api/workspaces/session1/canvases')
-      const listBody = (await listRes.json()) as {
-        canvases: { path: string; kind: string }[]
-      }
-      expect(listBody.canvases.find((c) => c.path === 'canvas-b')?.kind).toBe('markdown')
-    })
-
-    it('restoring into an existing target path WITHOUT overwrite returns 409 output_exists', async () => {
-      const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-
-      const sourceDoc = new LoroDoc()
-      const svv0 = sourceDoc.version()
-      const sourceList = sourceDoc.getMovableList('elements')
-      const sm0 = sourceList.insertContainer(0, new LoroMap())
-      sm0.set('id', 'keep-me')
-      sourceDoc.commit()
-      await app.request('/api/w/session1/canvas/canvas-a/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: sourceDoc.export({ mode: 'update', from: svv0 }),
-      })
-      const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ label: 'v1' }),
-      })
-      const saveBody = (await saveRes.json()) as { version: { id: string } }
-
-      await app.request('/api/w/session1/canvas/canvas-b/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/octet-stream' },
-        body: sourceDoc.export({ mode: 'update', from: svv0 }),
-      })
-
-      const restoreRes = await app.request(
-        `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/restore`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ targetPath: 'canvas-b' }),
-        },
-      )
-      expect(restoreRes.status).toBe(409)
-      const body = (await restoreRes.json()) as { error: string }
-      expect(body.error).toBe('output_exists')
-    })
-  })
-
-  // Thumbnail PUT/GET endpoint coverage.
-  it('saves a PNG through PUT /versions/:id/thumbnail and fetches it through GET', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-    const id = saveBody.version.id
-
-    // Minimal bytes starting with the PNG signature.
-    const png = new Uint8Array([0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a])
-    const putRes = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${id}/thumbnail`,
-      {
-        method: 'PUT',
-        headers: { 'Content-Type': 'image/png' },
-        body: png,
-      },
-    )
-    expect(putRes.status).toBe(200)
-
-    const getRes = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${id}/thumbnail`,
-    )
-    expect(getRes.status).toBe(200)
-    expect(getRes.headers.get('content-type')).toBe('image/png')
-    const bytes = new Uint8Array(await getRes.arrayBuffer())
-    expect(Array.from(bytes)).toEqual(Array.from(png))
-  })
-
-  it('rejects an oversized thumbnail body with 413 payload_too_large', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-    const id = saveBody.version.id
-
-    const oversized = new Uint8Array(16 * 1024 * 1024 + 1)
-    const putRes = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${id}/thumbnail`,
-      {
-        method: 'PUT',
-        headers: { 'Content-Type': 'image/png' },
-        body: oversized,
-      },
-    )
-    expect(putRes.status).toBe(413)
-    const body: unknown = await putRes.json()
-    expect(body).toMatchObject({ error: 'payload_too_large' })
-  })
-
-  it('rejects non-PNG magic bytes such as JPEG with 400 on PUT /thumbnail', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-    // JPEG signature (FF D8 FF)
-    const notPng = new Uint8Array([0xff, 0xd8, 0xff, 0xe0, 0x00, 0x10, 0x4a, 0x46])
-    const res = await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/thumbnail`,
-      { method: 'PUT', headers: { 'Content-Type': 'image/jpeg' }, body: notPng },
-    )
-    expect(res.status).toBe(400)
-  })
-
-  it('returns 404 for an unsaved thumbnail id on GET /thumbnail', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request(
-      '/api/workspaces/session1/canvases/canvas-a/versions/no-thumb/thumbnail',
-    )
-    expect(res.status).toBe(404)
-  })
-
-  // GET /latest-thumbnail is consumed by CanvasThumb's <img src>. A 404 makes
-  // the browser log "Failed to load resource: 404" for every thumbnail-less
-  // canvas, which is console noise (the component already has an onError
-  // fallback). Returning 204 No Content is a success status, so no console
-  // error is logged, while the empty body still triggers <img> onError → the
-  // FileText placeholder.
-  it('returns 204 (not 404) from GET /latest-thumbnail when the canvas has no thumbnail', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a/latest-thumbnail')
-    expect(res.status).toBe(204)
-    expect(await res.arrayBuffer()).toEqual(new ArrayBuffer(0))
-  })
-
-  it('returns 204 from GET /latest-thumbnail when metadata claims a thumbnail but the blob is gone', async () => {
-    // A version is listed as hasThumbnail=true but loadThumbnail resolves null
-    // (blob pruned/missing without throwing). This is the second 204 branch,
-    // distinct from "no thumbnailed version exists": the <img> still needs a
-    // success status to avoid 404 console noise.
-    const versionStore = {
-      save: vi.fn(),
-      load: vi.fn(),
-      list: vi.fn().mockResolvedValue([{ id: 'v1', hasThumbnail: true }]),
-      saveThumbnail: vi.fn(),
-      loadThumbnail: vi.fn().mockResolvedValue(null),
-      earliestFrontiers: vi.fn().mockResolvedValue([]),
-      getFrontiersBase64: vi.fn(),
-      renameBranchInVersions: vi.fn(),
-      pruneSandwichedAutoVersions: vi.fn().mockResolvedValue({ deletedCount: 0, deletedIds: [] }),
-    } as unknown as Parameters<typeof createCanvasRouter>[0]['versionStore']
-
-    const app = createCanvasRouter({ versionStore })
-    const res = await app.request('/api/workspaces/session1/canvases/canvas-a/latest-thumbnail')
-    expect(res.status).toBe(204)
-    expect(await res.arrayBuffer()).toEqual(new ArrayBuffer(0))
-  })
-
-  it('returns structured 500 for a broken thumbnail file on GET /thumbnail', async () => {
-    await mkdir(join(tmp.dir, 'blobs', 'session1', 'versions', 'broken-thumb.png'), {
-      recursive: true,
-    })
-
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const res = await app.request(
-      '/api/workspaces/session1/canvases/canvas-a/versions/broken-thumb/thumbnail',
-    )
-
-    expect(res.status).toBe(500)
-    await expect(res.json()).resolves.toEqual({
-      error: 'corrupt_stored_data',
-      message: expect.stringContaining('broken-thumb.png'),
-    })
-  })
-
-  it('returns hasThumbnail=true in the version list after saving a thumbnail', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    const saveRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ label: 'v1' }),
-    })
-    const saveBody = (await saveRes.json()) as { version: { id: string } }
-    const png = new Uint8Array([0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a])
-    await app.request(
-      `/api/workspaces/session1/canvases/canvas-a/versions/${saveBody.version.id}/thumbnail`,
-      { method: 'PUT', headers: { 'Content-Type': 'image/png' }, body: png },
-    )
-    const listRes = await app.request('/api/workspaces/session1/canvases/canvas-a/versions')
-    const listBody = (await listRes.json()) as { versions: Array<{ hasThumbnail: boolean }> }
-    expect(listBody.versions[0].hasThumbnail).toBe(true)
-  })
-
-  it('filters GET /versions by path and returns newest first', async () => {
-    const app = createCanvasRouter({ autoVersionIntervalMs: 60_000 })
-    await app.request('/api/workspaces/session1/canvases/canvas-a/versions', {
-      method: 'POST',
-      body: JSON.stringify({ label: 'a1' }),
-      headers: { 'Content-Type': 'application/json' },
-    })
-    await app.request('/api/workspaces/session1/canvases/canvas-b/versions', {
-      method: 'POST',
-      body: JSON.stringify({ label: 'b1' }),
-      headers: { 'Content-Type': 'application/json' },
-    })
-
-    const resA = await app.request('/api/workspaces/session1/canvases/canvas-a/versions')
-    const bodyA = (await resA.json()) as { versions: Array<{ label?: string }> }
-    expect(bodyA.versions.map((v) => v.label)).toEqual(['a1'])
-
-    const resB = await app.request('/api/workspaces/session1/canvases/canvas-b/versions')
-    const bodyB = (await resB.json()) as { versions: Array<{ label?: string }> }
-    expect(bodyB.versions.map((v) => v.label)).toEqual(['b1'])
-  })
-})
-
-describe('createAutoVersionTrigger', () => {
-  // One test in this block pins the clock with fake timers; always restore real timers
-  // after each test so a failed assertion can't leak a frozen clock into later tests.
-  afterEach(() => {
-    vi.useRealTimers()
-  })
-
-  it('retries on the next edit without consuming the throttle window when save fails', async () => {
-    const doc = new LoroDoc()
-    const entry = {
-      id: 'v1',
-      path: 'canvas-a',
-      createdAt: '2026-04-23T00:00:00.000Z',
-      elementCount: 0,
-      auto: true,
-      hasThumbnail: false,
-    }
-    const save = vi
-      .fn()
-      .mockRejectedValueOnce(new Error('transient fs error'))
-      .mockResolvedValueOnce(entry)
-    const trigger = createAutoVersionTrigger(
-      {
-        save,
-        load: vi.fn(),
-        list: vi.fn(),
-        saveThumbnail: vi.fn(),
-        loadThumbnail: vi.fn(),
-        earliestFrontiers: vi.fn(),
-        getFrontiersBase64: vi.fn(),
-      },
-      30_000,
-    )
-
-    await expect(trigger('session1', 'canvas-a', doc)).resolves.toBeNull()
-    await expect(trigger('session1', 'canvas-a', doc)).resolves.toEqual(entry)
-    expect(save).toHaveBeenCalledTimes(2)
-  })
-
-  it('passes branchName to save when getHeadBranch is injected', async () => {
-    const doc = new LoroDoc()
-    const entry = {
-      id: 'v1',
-      path: 'canvas-a',
-      createdAt: '2026-04-23T00:00:00.000Z',
-      elementCount: 0,
-      auto: true,
-      hasThumbnail: false,
-      branchName: 'feature',
-    }
-    const save = vi.fn().mockResolvedValue(entry)
-    const getHeadBranch = vi
-      .fn<(sid: string, path: string) => Promise<string | null>>()
-      .mockResolvedValue('feature')
-    const trigger = createAutoVersionTrigger(
-      {
-        save,
-        load: vi.fn(),
-        list: vi.fn(),
-        saveThumbnail: vi.fn(),
-        loadThumbnail: vi.fn(),
-        earliestFrontiers: vi.fn(),
-        getFrontiersBase64: vi.fn(),
-        renameBranchInVersions: vi.fn(),
-      },
-      30_000,
-      getHeadBranch,
-    )
-
-    await trigger('session1', 'canvas-a', doc)
-    expect(getHeadBranch).toHaveBeenCalledWith('session1', 'canvas-a')
-    expect(save).toHaveBeenCalledWith('session1', 'canvas-a', doc, {
-      auto: true,
-      branchName: 'feature',
-      operator: {
-        kind: 'system',
-        peerId: doc.peerIdStr,
-        displayName: 'auto-save',
-      },
-    })
-  })
-
-  it('calls save without branchName when getHeadBranch returns null', async () => {
-    const doc = new LoroDoc()
-    const save = vi.fn().mockResolvedValue({
-      id: 'v1',
-      path: 'canvas-a',
-      createdAt: '2026-04-23T00:00:00.000Z',
-      elementCount: 0,
-      auto: true,
-      hasThumbnail: false,
-      branchName: 'main',
-    })
-    const getHeadBranch = vi.fn().mockResolvedValue(null)
-    const trigger = createAutoVersionTrigger(
-      {
-        save,
-        load: vi.fn(),
-        list: vi.fn(),
-        saveThumbnail: vi.fn(),
-        loadThumbnail: vi.fn(),
-        earliestFrontiers: vi.fn(),
-        getFrontiersBase64: vi.fn(),
-        renameBranchInVersions: vi.fn(),
-      },
-      30_000,
-      getHeadBranch,
-    )
-    await trigger('session1', 'canvas-a', doc)
-    expect(save).toHaveBeenCalledWith('session1', 'canvas-a', doc, {
-      auto: true,
-      operator: {
-        kind: 'system',
-        peerId: doc.peerIdStr,
-        displayName: 'auto-save',
-      },
-    })
-  })
-
-  it('does not silently fall back to save when getHeadBranch throws corruption', async () => {
-    const doc = new LoroDoc()
-    const save = vi.fn()
-    const getHeadBranch = vi
-      .fn<(sid: string, path: string) => Promise<string | null>>()
-      .mockRejectedValue(corruptStoredData('/tmp/branches.json', 'broken branch state'))
-    const trigger = createAutoVersionTrigger(
-      {
-        save,
-        load: vi.fn(),
-        list: vi.fn(),
-        saveThumbnail: vi.fn(),
-        loadThumbnail: vi.fn(),
-        earliestFrontiers: vi.fn(),
-        getFrontiersBase64: vi.fn(),
-        renameBranchInVersions: vi.fn(),
-      },
-      30_000,
-      getHeadBranch,
-    )
-
-    await expect(trigger('session1', 'canvas-a', doc)).rejects.toMatchObject({
-      name: 'CorruptStoredDataError',
-      message: expect.stringContaining('broken branch state'),
-    })
-    expect(save).not.toHaveBeenCalled()
-  })
-
-  it('returns null without consuming the throttle window when versionStore.save throws corruption', async () => {
-    const doc = new LoroDoc()
-    const entry = {
-      id: 'v2',
-      path: 'canvas-a',
-      createdAt: '2026-04-23T00:00:00.000Z',
-      elementCount: 0,
-      auto: true,
-      hasThumbnail: false,
-      branchName: 'main',
-    }
-    const save = vi
-      .fn()
-      .mockRejectedValueOnce(corruptStoredData('/tmp/versions/v1.json', 'broken metadata'))
-      .mockResolvedValueOnce(entry)
-    const trigger = createAutoVersionTrigger(
-      {
-        save,
-        load: vi.fn(),
-        list: vi.fn(),
-        saveThumbnail: vi.fn(),
-        loadThumbnail: vi.fn(),
-        earliestFrontiers: vi.fn(),
-        getFrontiersBase64: vi.fn(),
-        renameBranchInVersions: vi.fn(),
-      },
-      30_000,
-    )
-
-    await expect(trigger('session1', 'canvas-a', doc)).resolves.toBeNull()
-    await expect(trigger('session1', 'canvas-a', doc)).resolves.toEqual(entry)
-    expect(save).toHaveBeenCalledTimes(2)
-  })
-
-  it('throttles repeated calls within the interval window by tracking last-save time per canvas key', async () => {
-    // The trigger holds an ephemeral per-canvas timestamp registry (Map<key, number>).
-    // This test confirms the registry persists its state across calls within the same
-    // trigger instance so that the throttle window is correctly enforced.
-    //
-    // Pin the clock so both the "first call always fires" invariant and the
-    // "second call is throttled" assertion are grounded in an explicit time,
-    // not a wall-clock assumption.
-    vi.useFakeTimers()
-    vi.setSystemTime(new Date('2026-01-01T00:00:00.000Z'))
-
-    const doc = new LoroDoc()
-    const entry = {
-      id: 'v1',
-      path: 'canvas-a',
-      createdAt: '2026-04-23T00:00:00.000Z',
-      elementCount: 0,
-      auto: true,
-      hasThumbnail: false,
-    }
-    const save = vi.fn().mockResolvedValue(entry)
-    // Use a large interval so the second call (at the same pinned instant) always falls within the window.
-    const trigger = createAutoVersionTrigger(
-      {
-        save,
-        load: vi.fn(),
-        list: vi.fn(),
-        saveThumbnail: vi.fn(),
-        loadThumbnail: vi.fn(),
-        earliestFrontiers: vi.fn(),
-        getFrontiersBase64: vi.fn(),
-      },
-      60_000,
-    )
-
-    // First call: no prior save recorded → now - 0 >= intervalMs is true → should save.
-    const first = await trigger('session1', 'canvas-a', doc)
-    expect(first).toEqual(entry)
-    expect(save).toHaveBeenCalledTimes(1)
-
-    // Second call at the same pinned instant: within the 60s window → must return null.
-    const second = await trigger('session1', 'canvas-a', doc)
-    expect(second).toBeNull()
-    expect(save).toHaveBeenCalledTimes(1)
-
-    vi.useRealTimers()
-  })
-})
-
-describe('auto-version corruption handling', () => {
-  beforeEach(async () => {
-    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
-    clearCache()
-  })
-  afterEach(() => {
-    clearCache()
-  })
-
-  it('returns 200 and skips version_created when auto-version save reports corruption', async () => {
-    const wsModule = await import('./ws.js')
-    const sendVersionCreated = vi
-      .spyOn(wsModule, 'sendVersionCreated')
-      .mockImplementation(() => undefined)
-    const versionStore = {
-      save: vi
-        .fn()
-        .mockRejectedValue(corruptStoredData('/tmp/versions/v1.json', 'broken metadata')),
-      load: vi.fn(),
-      list: vi.fn(),
-      saveThumbnail: vi.fn(),
-      loadThumbnail: vi.fn(),
-      earliestFrontiers: vi.fn(),
-      getFrontiersBase64: vi.fn(),
-      renameBranchInVersions: vi.fn(),
-    }
-
-    const clientDoc = new LoroDoc()
-    const prevVV = clientDoc.version()
-    const list = clientDoc.getMovableList('elements')
-    const map = list.insertContainer(0, new LoroMap())
-    map.set('id', 'e1')
-    map.set('type', 'rectangle')
-    clientDoc.commit()
-    const update = clientDoc.export({ mode: 'update', from: prevVV })
-
-    const app = createCanvasRouter({
-      autoVersionIntervalMs: 0,
-      versionStore,
-    })
-    const res = await app.request('/api/w/session1/canvas/canvas-a/update', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: update,
-    })
-
-    expect(res.status).toBe(200)
-    await new Promise((resolve) => setTimeout(resolve, 50))
-    expect(versionStore.save).toHaveBeenCalledTimes(1)
-    expect(sendVersionCreated).not.toHaveBeenCalled()
-
-    clearCache()
-    const { loadDocument } = await import('../store/document-store.js')
-    const serverDoc = await loadDocument('session1', 'canvas-a')
-    const elements = serverDoc.getMovableList('elements').toJSON() as Array<{ id: string }>
-    expect(elements.map((entry) => entry.id)).toEqual(['e1'])
   })
 })
