+++ packages/mcp-server/src/server/routes/canvas/workspaces.test.ts
@@ -1,20 +1,733 @@
+import { mkdir, writeFile } from 'node:fs/promises'
+import { join } from 'node:path'
 import { Hono } from 'hono'
-import { describe, expect, it, vi } from 'vitest'
+import { LoroDoc, LoroMap } from 'loro-crdt'
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
+import {
+  deleteCanvasResponseSchema,
+  renameDocumentPathResponseSchema,
+} from '../../../shared/api-contracts/canvas.js'
+import { withTempDataDir } from '../_test-helpers.js'
+
+const tmp = withTempDataDir('whiteboard-workspaces-test-')
 
 vi.mock('../../config.js', () => ({
   get DATA_DIR() {
-    return '/tmp/test-workspaces'
+    return tmp.dir
   },
-  getDataDir: () => '/tmp/test-workspaces',
+  getDataDir: () => tmp.dir,
   WHITEBOARD_ROOT: '/tmp/whiteboard',
   REPO_ROOT: '/tmp',
 }))
 
+// Mock doc-cache so the cache is isolated in tests. getDoc is wrapped in
+// vi.fn (defaulting to the real implementation) so individual tests can
+// install a one-shot mockImplementationOnce to control interleaving against
+// a concurrent request.
+vi.mock('../../store/doc-cache.js', async () => {
+  const actual = await vi.importActual<typeof import('../../store/doc-cache.js')>(
+    '../../store/doc-cache.js',
+  )
+  return { ...actual, getDoc: vi.fn(actual.getDoc) }
+})
+
+const { clearCache, peekDoc, getDoc } = await import('../../store/doc-cache.js')
+const { saveDocument } = await import('../../store/document-store.js')
+const { corruptStoredData } = await import('../../store/corrupt-stored-data.js')
 const { createWorkspacesRouter } = await import('./workspaces.js')
+const { createCanvasRouter } = await import('../canvas.js')
 
 describe('workspaces router', () => {
   it('returns a Hono instance', () => {
     const app = createWorkspacesRouter()
     expect(app).toBeInstanceOf(Hono)
+  })
+})
+
+describe('GET /api/workspaces', () => {
+  beforeEach(() => {
+    clearCache()
+  })
+  afterEach(() => {
+    clearCache()
+  })
+
+  it('returns the canonical workspace list with workspaceId entries', async () => {
+    await saveDocument('workspace-a', 'a', new LoroDoc())
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces')
+
+    expect(res.status).toBe(200)
+    const json = (await res.json()) as {
+      workspaces: Array<{ workspaceId: string }>
+    }
+    expect(json.workspaces).toEqual([{ workspaceId: 'workspace-a' }])
+  })
+})
+
+describe('POST /api/workspaces/:workspaceId/canvases', () => {
+  beforeEach(() => {
+    clearCache()
+  })
+  afterEach(() => {
+    clearCache()
+  })
+
+  it('returns { path } on success', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'new-canvas' }),
+    })
+    expect(res.status).toBe(200)
+    const json = (await res.json()) as unknown
+    expect(json).toEqual({ path: 'new-canvas' })
+  })
+
+  it('returns 409 with Problem Details title on duplicate path', async () => {
+    const app = createCanvasRouter()
+    // Create once to seed the conflict.
+    await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'existing' }),
+    })
+    // Second creation must return Problem Details with a title field.
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'existing' }),
+    })
+    expect(res.status).toBe(409)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    expect(json.title!.length).toBeGreaterThan(0)
+  })
+
+  it('returns 500 with Problem Details title when saveDocument fails unexpectedly', async () => {
+    // Block the canvas blob directory with a file so saveDocument throws a
+    // non-ConflictError, exercising the catch-all 500 branch (mutation-check
+    // guard for the 400 -> 500 change).
+    await mkdir(join(tmp.dir, 'blobs', 'ws1'), { recursive: true })
+    await writeFile(join(tmp.dir, 'blobs', 'ws1', 'canvas'), 'not-a-directory')
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'new-canvas' }),
+    })
+    expect(res.status).toBe(500)
+    const json = (await res.json()) as { title?: string }
+    expect(json.title).toBe('Failed to create canvas.')
+  })
+
+  it('returns 400 with Problem Details title on invalid path', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'bad path!' }),
+    })
+    expect(res.status).toBe(400)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+  })
+
+  it('returns 400 with Problem Details title when body is not JSON', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'text/plain' },
+      body: 'not-json',
+    })
+    expect(res.status).toBe(400)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    expect(json.title!.length).toBeGreaterThan(0)
+  })
+
+  it('returns 400 with Problem Details title when body has no path field', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ name: 'oops' }),
+    })
+    expect(res.status).toBe(400)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    expect(json.title!.length).toBeGreaterThan(0)
+  })
+
+  it('creates a kind:markdown canvas, and the list carries it back', async () => {
+    const app = createCanvasRouter()
+    const createRes = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'note', kind: 'markdown' }),
+    })
+    expect(createRes.status).toBe(200)
+    // Response body stays exactly { path } — kind is not echoed back.
+    expect(await createRes.json()).toEqual({ path: 'note' })
+
+    const listRes = await app.request('/api/workspaces/ws1/canvases')
+    const listJson = (await listRes.json()) as { canvases: { path: string; kind: string }[] }
+    const created = listJson.canvases.find((c) => c.path === 'note')
+    expect(created?.kind).toBe('markdown')
+
+    // The empty LoroDoc a markdown-kind create saves loads without error —
+    // an empty doc is a valid initial document for either kind.
+    const snapshotRes = await app.request('/api/w/ws1/canvas/note/snapshot')
+    expect(snapshotRes.status).toBe(200)
+  })
+
+  it('creates a canvas without kind — response and list stay byte-identical to spatial back-compat', async () => {
+    const app = createCanvasRouter()
+    const createRes = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'legacy' }),
+    })
+    expect(createRes.status).toBe(200)
+    expect(await createRes.json()).toEqual({ path: 'legacy' })
+
+    const listRes = await app.request('/api/workspaces/ws1/canvases')
+    const listJson = (await listRes.json()) as { canvases: { path: string; kind: string }[] }
+    expect(listJson.canvases.find((c) => c.path === 'legacy')?.kind).toBe('spatial')
+  })
+
+  it('returns 400 with Problem Details title naming the actual failing field on an invalid kind', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'bad-kind', kind: 'bogus' }),
+    })
+    expect(res.status).toBe(400)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    // A valid path plus an invalid kind must not be told "path is required" —
+    // that names the wrong field and gives no path to recovery.
+    expect(json.title).toMatch(/kind/i)
+    expect(json.title).not.toMatch(/path is required/i)
+  })
+
+  it('falls back to the issue message when the invalid field is neither path nor kind', async () => {
+    // An array body parses as JSON but fails the object schema at the TOP level — path [] —
+    // exercising createCanvasRequestErrorTitle's fallback branch rather than a field-specific
+    // message that would name the wrong thing.
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/ws-a/canvases', {
+      method: 'POST',
+      headers: { 'content-type': 'application/json' },
+      body: JSON.stringify([1, 2, 3]),
+    })
+    expect(res.status).toBe(400)
+    const json = (await res.json()) as { title: string }
+    expect(json.title).not.toMatch(/path is required/i)
+    expect(json.title).not.toMatch(/spatial/i)
+    expect(json.title.length).toBeGreaterThan(0)
+  })
+
+  it('returns 400 with Problem Details { title } (not legacy { error, message }) on invalid workspaceId', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/bad.workspace/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'test' }),
+    })
+    expect(res.status).toBe(400)
+    const json = (await res.json()) as Record<string, unknown>
+    // Must have a Problem Details title, not the legacy { error, message } shape.
+    expect(typeof json.title).toBe('string')
+    expect(json.title as string).toBeTruthy()
+    // Must NOT carry the old shape keys.
+    expect(json).not.toHaveProperty('error')
+    expect(json).not.toHaveProperty('message')
+  })
+})
+
+describe('DELETE /api/workspaces/:workspaceId/canvases/:path', () => {
+  beforeEach(async () => {
+    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
+    clearCache()
+  })
+  afterEach(() => {
+    clearCache()
+  })
+
+  it('returns 200 { ok: true }, parses with deleteCanvasResponseSchema, and the canvas is gone from list/exists/snapshot', async () => {
+    await saveDocument('session1', 'canvas-a', new LoroDoc())
+    const app = createCanvasRouter()
+
+    const res = await app.request('/api/workspaces/session1/canvases/canvas-a', {
+      method: 'DELETE',
+    })
+    expect(res.status).toBe(200)
+    const json: unknown = await res.json()
+    expect(deleteCanvasResponseSchema.parse(json)).toEqual({ ok: true })
+
+    const listRes = await app.request('/api/workspaces/session1/canvases')
+    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
+    expect(listJson.canvases.map((c) => c.path)).not.toContain('canvas-a')
+
+    const existsRes = await app.request('/api/w/session1/canvas/canvas-a/exists')
+    expect(await existsRes.json()).toEqual({ exists: false })
+
+    const snapshotRes = await app.request('/api/w/session1/canvas/canvas-a/snapshot')
+    expect(snapshotRes.status).toBe(404)
+  })
+
+  it('returns 404 with Problem Details { title } for a missing canvas', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/session1/canvases/never-created', {
+      method: 'DELETE',
+    })
+    expect(res.status).toBe(404)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    expect(json.title!.length).toBeGreaterThan(0)
+    expect(json).not.toHaveProperty('error')
+  })
+
+  it('returns 400 with Problem Details { title } for an invalid workspaceId or path', async () => {
+    const app = createCanvasRouter()
+
+    const badWs = await app.request('/api/workspaces/bad.workspace/canvases/canvas-a', {
+      method: 'DELETE',
+    })
+    expect(badWs.status).toBe(400)
+    expect(typeof ((await badWs.json()) as { title?: string }).title).toBe('string')
+
+    const badPath = await app.request('/api/workspaces/session1/canvases/bad.path', {
+      method: 'DELETE',
+    })
+    expect(badPath.status).toBe(400)
+    expect(typeof ((await badPath.json()) as { title?: string }).title).toBe('string')
+  })
+
+  it('maps a thrown CorruptStoredDataError to 500 { error: corrupt_stored_data }, and only that error type — a plain throw stays a generic 500', async () => {
+    await saveDocument('session1', 'canvas-a', new LoroDoc())
+    const canvasStore = await import('../../store/document-store.js')
+    const spy = vi
+      .spyOn(canvasStore, 'deleteDocument')
+      .mockRejectedValueOnce(
+        corruptStoredData('/tmp/blobs/session1/canvas/abc.loro', 'broken canvas blob'),
+      )
+    const app = createCanvasRouter()
+    try {
+      const res = await app.request('/api/workspaces/session1/canvases/canvas-a', {
+        method: 'DELETE',
+      })
+      expect(res.status).toBe(500)
+      await expect(res.json()).resolves.toEqual({
+        error: 'corrupt_stored_data',
+        message: expect.stringContaining('broken canvas blob'),
+      })
+
+      // Mutation guard: a non-corruption throw must NOT hit the same
+      // branch — proves the mapping checks the error type, not "any throw".
+      spy.mockRejectedValueOnce(new Error('disk exploded'))
+      const res2 = await app.request('/api/workspaces/session1/canvases/canvas-a', {
+        method: 'DELETE',
+      })
+      expect(res2.status).toBe(500)
+      const json2: unknown = await res2.json()
+      expect(json2).not.toMatchObject({ error: 'corrupt_stored_data' })
+    } finally {
+      spy.mockRestore()
+    }
+  })
+
+  it('evicts the doc-cache on delete, so re-creating the same path after a warm cache read yields a fresh empty doc', async () => {
+    const app = createCanvasRouter()
+    await app.request('/api/workspaces/session1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'cached' }),
+    })
+    const clientDoc = new LoroDoc()
+    const prevVV = clientDoc.version()
+    const list = clientDoc.getMovableList('elements')
+    const m = list.insertContainer(0, new LoroMap())
+    m.set('id', 'old-element')
+    clientDoc.commit()
+    const update = clientDoc.export({ mode: 'update', from: prevVV })
+    // Warm the doc-cache via the update path, matching real client traffic.
+    await app.request('/api/w/session1/canvas/cached/update', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/octet-stream' },
+      body: update,
+    })
+    expect(peekDoc('session1', 'cached')).toBeDefined()
+
+    const delRes = await app.request('/api/workspaces/session1/canvases/cached', {
+      method: 'DELETE',
+    })
+    expect(delRes.status).toBe(200)
+
+    // Re-creating must succeed (not 409) and must not resurrect the old doc.
+    const createRes = await app.request('/api/workspaces/session1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'cached' }),
+    })
+    expect(createRes.status).toBe(200)
+
+    const snapshotRes = await app.request('/api/w/session1/canvas/cached/snapshot')
+    expect(snapshotRes.status).toBe(200)
+    const buf = await snapshotRes.arrayBuffer()
+    const restored = LoroDoc.fromSnapshot(new Uint8Array(buf))
+    expect(restored.getMovableList('elements').length).toBe(0)
+  })
+})
+
+describe('PUT /api/workspaces/:workspaceId/canvases/:path/path', () => {
+  beforeEach(async () => {
+    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
+    clearCache()
+  })
+  afterEach(() => {
+    clearCache()
+  })
+
+  it('returns 200 { path }, the list shows the new path and not the old one, the old snapshot URL 404s, and re-creating the old path afterward succeeds as a fresh canvas', async () => {
+    await saveDocument('session1', 'a', new LoroDoc())
+    const app = createCanvasRouter()
+
+    const res = await app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'b' }),
+    })
+    expect(res.status).toBe(200)
+    const json: unknown = await res.json()
+    expect(renameDocumentPathResponseSchema.parse(json)).toEqual({ path: 'b' })
+
+    const listRes = await app.request('/api/workspaces/session1/canvases')
+    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
+    const paths = listJson.canvases.map((c) => c.path)
+    expect(paths).toContain('b')
+    expect(paths).not.toContain('a')
+
+    const oldSnapshotRes = await app.request('/api/w/session1/canvas/a/snapshot')
+    expect(oldSnapshotRes.status).toBe(404)
+
+    const recreateRes = await app.request('/api/workspaces/session1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'a' }),
+    })
+    expect(recreateRes.status).toBe(200)
+  })
+
+  it('returns 404 with Problem Details { title } for a missing source canvas', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/session1/canvases/never-created/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'target' }),
+    })
+    expect(res.status).toBe(404)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    expect(json.title!.length).toBeGreaterThan(0)
+    expect(json).not.toHaveProperty('error')
+  })
+
+  it('returns 409 with Problem Details { title } when the target path is already taken', async () => {
+    await saveDocument('session1', 'a', new LoroDoc())
+    await saveDocument('session1', 'b', new LoroDoc())
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'b' }),
+    })
+    expect(res.status).toBe(409)
+    const json = (await res.json()) as { title?: string }
+    expect(typeof json.title).toBe('string')
+    expect(json.title!.length).toBeGreaterThan(0)
+
+    // Neither canvas was mutated by the rejected rename.
+    const listRes = await app.request('/api/workspaces/session1/canvases')
+    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
+    expect(listJson.canvases.map((c) => c.path).sort()).toEqual(['a', 'b'])
+  })
+
+  it('does not fork a phantom duplicate canvas when a rename races an in-flight /update that already resolved a doc reference through the old path', async () => {
+    const app = createCanvasRouter()
+    const baseDoc = new LoroDoc()
+    baseDoc.getText('content').insert(0, 'original')
+    baseDoc.commit()
+    await saveDocument('session1', 'a', baseDoc)
+
+    // A client update built against the pre-rename base.
+    const clientDoc = LoroDoc.fromSnapshot(baseDoc.export({ mode: 'snapshot' }))
+    const fromVV = clientDoc.version()
+    clientDoc.getText('content').insert('original'.length, ' + edit')
+    clientDoc.commit()
+    const update = clientDoc.export({ mode: 'update', from: fromVV })
+
+    // Stall the /update route's getDoc() call so a rename can be fired
+    // while the request is paused mid-flight, matching the real race: the
+    // read resolves before the rename runs, the write happens after.
+    let releaseGetDoc: () => void = () => undefined
+    const getDocGate = new Promise<void>((resolve) => {
+      releaseGetDoc = resolve
+    })
+    let signalGetDocCalled: () => void = () => undefined
+    const getDocCalled = new Promise<void>((resolve) => {
+      signalGetDocCalled = resolve
+    })
+    const actual = await vi.importActual<typeof import('../../store/doc-cache.js')>(
+      '../../store/doc-cache.js',
+    )
+    vi.mocked(getDoc).mockImplementationOnce(async (workspaceId, path) => {
+      signalGetDocCalled()
+      await getDocGate
+      return actual.getDoc(workspaceId, path)
+    })
+
+    const updatePromise = app.request('/api/w/session1/canvas/a/update', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/octet-stream' },
+      body: update,
+    })
+
+    await getDocCalled
+
+    // Fire the rename while the update is stalled mid-flight.
+    const renamePromise = app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'b' }),
+    })
+    // Give the rename a chance to run before letting the stalled read continue.
+    await new Promise((r) => setTimeout(r, 20))
+    releaseGetDoc()
+
+    const [updateRes, renameRes] = await Promise.all([updatePromise, renamePromise])
+    expect(renameRes.status).toBe(200)
+    expect(updateRes.status).toBe(200)
+
+    // Exactly one canvas must survive -- the update must not have silently
+    // inserted a phantom duplicate back at the old path.
+    const listRes = await app.request('/api/workspaces/session1/canvases')
+    const listJson = (await listRes.json()) as { canvases: { path: string }[] }
+    expect(listJson.canvases.map((c) => c.path)).toEqual(['b'])
+  })
+
+  it('returns 400 with Problem Details { title } for invalid input shapes', async () => {
+    await saveDocument('session1', 'a', new LoroDoc())
+    const app = createCanvasRouter()
+
+    const noBody = await app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+    })
+    expect(noBody.status).toBe(400)
+
+    const badNewPath = await app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'bad.path' }),
+    })
+    expect(badNewPath.status).toBe(400)
+
+    const emptyPath = await app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: '' }),
+    })
+    expect(emptyPath.status).toBe(400)
+
+    const badWorkspace = await app.request('/api/workspaces/bad.workspace/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'b' }),
+    })
+    expect(badWorkspace.status).toBe(400)
+
+    const badCurrentPath = await app.request('/api/workspaces/session1/canvases/bad.path/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'b' }),
+    })
+    expect(badCurrentPath.status).toBe(400)
+
+    for (const res of [noBody, badNewPath, emptyPath, badWorkspace, badCurrentPath]) {
+      const body = (await res.json()) as { title?: string }
+      expect(typeof body.title).toBe('string')
+      expect(body.title!.length).toBeGreaterThan(0)
+    }
+  })
+
+  it('evicts the doc-cache on rename: updating via the OLD path afterward lazily creates a FRESH canvas rather than resurrecting the warmed doc, while the new path keeps the real content', async () => {
+    const app = createCanvasRouter()
+    await app.request('/api/workspaces/session1/canvases', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'a' }),
+    })
+    const clientDoc = new LoroDoc()
+    const prevVV = clientDoc.version()
+    const list = clientDoc.getMovableList('elements')
+    const m = list.insertContainer(0, new LoroMap())
+    m.set('id', 'old-element')
+    clientDoc.commit()
+    const update = clientDoc.export({ mode: 'update', from: prevVV })
+    // Warm the doc-cache via the update path, matching real client traffic.
+    await app.request('/api/w/session1/canvas/a/update', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/octet-stream' },
+      body: update,
+    })
+    expect(peekDoc('session1', 'a')).toBeDefined()
+
+    const renameRes = await app.request('/api/workspaces/session1/canvases/a/path', {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ path: 'b' }),
+    })
+    expect(renameRes.status).toBe(200)
+
+    // The renamed canvas keeps its content, now reachable at the new path.
+    const bSnapshotRes = await app.request('/api/w/session1/canvas/b/snapshot')
+    expect(bSnapshotRes.status).toBe(200)
+    const bBuf = await bSnapshotRes.arrayBuffer()
+    const bRestored = LoroDoc.fromSnapshot(new Uint8Array(bBuf))
+    expect(
+      (bRestored.getMovableList('elements').toJSON() as { id: string }[]).map((e) => e.id),
+    ).toEqual(['old-element'])
+
+    // Updating through the OLD path lazily creates a FRESH canvas — it must
+    // not resurrect the evicted doc's history.
+    const anotherClientDoc = new LoroDoc()
+    const anotherPrevVV = anotherClientDoc.version()
+    const anotherList = anotherClientDoc.getMovableList('elements')
+    const anotherM = anotherList.insertContainer(0, new LoroMap())
+    anotherM.set('id', 'fresh-element')
+    anotherClientDoc.commit()
+    const anotherUpdate = anotherClientDoc.export({ mode: 'update', from: anotherPrevVV })
+    await app.request('/api/w/session1/canvas/a/update', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/octet-stream' },
+      body: anotherUpdate,
+    })
+
+    const freshSnapshotRes = await app.request('/api/w/session1/canvas/a/snapshot')
+    expect(freshSnapshotRes.status).toBe(200)
+    const freshBuf = await freshSnapshotRes.arrayBuffer()
+    const freshRestored = LoroDoc.fromSnapshot(new Uint8Array(freshBuf))
+    const freshIds = (freshRestored.getMovableList('elements').toJSON() as { id: string }[]).map(
+      (e) => e.id,
+    )
+    expect(freshIds).toEqual(['fresh-element'])
+    expect(freshIds).not.toContain('old-element')
+  })
+
+  it('maps a thrown CorruptStoredDataError to 500 { error: corrupt_stored_data }, and only that error type — a plain throw stays a generic 500', async () => {
+    await saveDocument('session1', 'a', new LoroDoc())
+    const canvasStore = await import('../../store/document-store.js')
+    const spy = vi
+      .spyOn(canvasStore, 'renameDocumentPath')
+      .mockRejectedValueOnce(
+        corruptStoredData('/tmp/blobs/session1/canvas/abc.loro', 'broken canvas blob'),
+      )
+    const app = createCanvasRouter()
+    try {
+      const res = await app.request('/api/workspaces/session1/canvases/a/path', {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ path: 'b' }),
+      })
+      expect(res.status).toBe(500)
+      await expect(res.json()).resolves.toEqual({
+        error: 'corrupt_stored_data',
+        message: expect.stringContaining('broken canvas blob'),
+      })
+
+      // Mutation guard: a non-corruption throw must NOT hit the same
+      // branch — proves the mapping checks the error type, not "any throw".
+      spy.mockRejectedValueOnce(new Error('disk exploded'))
+      const res2 = await app.request('/api/workspaces/session1/canvases/a/path', {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ path: 'b' }),
+      })
+      expect(res2.status).toBe(500)
+      const json2: unknown = await res2.json()
+      expect(json2).not.toMatchObject({ error: 'corrupt_stored_data' })
+    } finally {
+      spy.mockRestore()
+    }
+  })
+})
+
+describe('GET /api/workspaces/:workspaceId/canvases', () => {
+  beforeEach(async () => {
+    await mkdir(join(tmp.dir, 'session1'), { recursive: true })
+    clearCache()
+  })
+  afterEach(() => {
+    clearCache()
+  })
+
+  it('returns the canvas list', async () => {
+    await saveDocument('session1', 'canvas-a', new LoroDoc())
+    await saveDocument('session1', 'canvas-b', new LoroDoc())
+
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/session1/canvases')
+    expect(res.status).toBe(200)
+    const json = (await res.json()) as { canvases: { path: string }[] }
+    const paths = json.canvases.map((c) => c.path)
+    expect(paths).toContain('canvas-a')
+    expect(paths).toContain('canvas-b')
+  })
+
+  it('returns 400 for an invalid workspaceId', async () => {
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/bad.sid/canvases')
+    expect(res.status).toBe(400)
+  })
+
+  it('returns 404 for a workspace that does not exist, not an empty list', async () => {
+    // A 200-empty here is a false statement of fact: the caller cannot tell
+    // "this workspace has no canvases" from "you asked about a workspace this
+    // daemon has never heard of", and the web app renders the first reading —
+    // an empty state with a Create button — for what is actually the second.
+    // A stale pairing (a workspace id minted by an earlier install and kept
+    // in the browser's localStorage) then looks exactly like data loss.
+    // The v1 document routes already 404 an unknown workspace; this brings
+    // the legacy list route into agreement.
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/never-registered/canvases')
+    expect(res.status).toBe(404)
+  })
+
+  // listDocuments no longer walks per-workspace directories, so the previous
+  // "broken session directory" 500 case no longer applies.
+})
+
+describe('GET /api/workspaces/:workspaceId/canvases', () => {
+  beforeEach(async () => {
+    await mkdir(join(tmp.dir, 'workspace1'), { recursive: true })
+    clearCache()
+  })
+  afterEach(() => {
+    clearCache()
+  })
+
+  it('also returns the canvas list from the canonical workspace route', async () => {
+    await saveDocument('workspace1', 'canvas-a', new LoroDoc())
+
+    const app = createCanvasRouter()
+    const res = await app.request('/api/workspaces/workspace1/canvases')
+
+    expect(res.status).toBe(200)
+    const json = (await res.json()) as { canvases: { path: string }[] }
+    expect(json.canvases).toEqual([expect.objectContaining({ path: 'canvas-a' })])
   })
 })
